# Conflicts:
#	Assets/Goat/_Scripts/Saving/SaveHandler.cs

Removed async function
Added coroutines to loading
Added csc file for removing the 0649 warning (created but never assigned warning)

+++ Assets/Goat/_Scripts/TimeManipulation/TimeSetter.cs
@@ -8,7 +8,7 @@ using UnityEngine.UI;
 public class TimeSetter : MonoBehaviour, IAtomListener<UnityAtoms.Void>
 {
     [SerializeField] private Button timeButton;
-    [SerializeField] private int timeScaleSet;
+    [SerializeField] private float timeScaleSet;
     [SerializeField] private int timeSpeedSet;
     [SerializeField] private IntEvent onTimeSpeedChanged;
     [SerializeField] private VoidEvent onTimeButtonClicked;
